# specs/012-planejamento-efetivo/contracts/efetivo-planning.openapi.yaml
openapi: 3.1.0
info:
  title: Efetivo Operacional Planning API
  version: 1.1.0
  description: Contrato das áreas de capacidade, calendário compartilhado, missões, execução observada, simulações e administração do Efetivo.
servers:
  - url: /api/efetivo
security:
  - sessionCookie: []
...
paths:
...
  /planning/collaborators:
    get:
      tags: [Collaborators]
      summary: Lista situação e projeção dos colaboradores na data
      parameters:
        - $ref: '#/components/parameters/Date'
        - $ref: '#/components/parameters/JobRoleId'
        - $ref: '#/components/parameters/Search'
      responses:
        '200':
          description: Colaboradores
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CollaboratorPlanning' }
    post:
      tags: [Collaborators]
      summary: Cria colaborador pelos campos operacionais permitidos
      x-required-role: efetivo:manager
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CollaboratorInput' }
      responses:
        '201': { $ref: '#/components/responses/CollaboratorResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
  /planning/collaborators/{collaboratorId}:
    patch:
      tags: [Collaborators]
      summary: Atualiza somente campos operacionais do cadastro canônico
      x-required-role: efetivo:manager
      parameters:
        - $ref: '#/components/parameters/CollaboratorId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CollaboratorInput' }
      responses:
        '200': { $ref: '#/components/responses/CollaboratorResponse' }
        '409': { $ref: '#/components/responses/ConflictResponse' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: connect.sid
  parameters:
    Date:
      name: date
      in: query
      required: true
      schema: { $ref: '#/components/schemas/DateOnly' }
...
    Search:
      name: search
      in: query
      schema: { type: string, maxLength: 120 }
    JobRoleId:
      name: jobRoleId
      in: query
      schema: { type: string }
...
    CollaboratorId:
      name: collaboratorId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    ValidationError:
      description: Entrada inválida
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    ConflictResponse:
      description: Conflito temporal, de revisão, vínculo, função ou demanda
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    CollaboratorResponse:
      description: Colaborador canônico salvo
      content:
        application/json:
          schema: { $ref: '#/components/schemas/CollaboratorPlanning' }
  schemas:
    DateOnly:
      type: string
      format: date
      pattern: '^\d{4}-\d{2}-\d{2}$'
...
    CollaboratorOption:
      type: object
      required: [id, name, jobRoleId, jobRoleName]
      properties:
        id: { type: string }
        name: { type: string }
        jobRoleId: { type: string }
        jobRoleName: { type: string }
    CollaboratorInput:
      type: object
      required: [name, jobRoleId, admissionDate]
      properties:
        name: { type: string, minLength: 1, maxLength: 160 }
        jobRoleId: { type: string }
        admissionDate: { $ref: '#/components/schemas/DateOnly' }
        terminationDate:
          oneOf:
            - { $ref: '#/components/schemas/DateOnly' }
            - { type: 'null' }
        note: { type: [string, 'null'], maxLength: 1000 }
    CollaboratorPlanning:
      allOf:
        - $ref: '#/components/schemas/CollaboratorOption'
        - type: object
          required: [admissionDate, status, plannedUtilization90d]
          properties:
            admissionDate: { $ref: '#/components/schemas/DateOnly' }
            terminationDate:
              oneOf:
                - { $ref: '#/components/schemas/DateOnly' }
                - { type: 'null' }
            status: { type: string, enum: [ALLOCATED, UNAVAILABLE, FREE, OUTSIDE_EMPLOYMENT] }
            plannedUtilization90d: { type: [number, 'null'], minimum: 0, maximum: 100 }
            vacationAlert: { type: [object, 'null'] }
...
    Conflict:
      type: object
      required: [code, collaboratorId, collaboratorName, startDate, endDate, sourceType, sourceId]
      properties:
        code: { type: string }
        collaboratorId: { type: string }
        collaboratorName: { type: string }
        startDate: { $ref: '#/components/schemas/DateOnly' }
        endDate: { $ref: '#/components/schemas/DateOnly' }
        sourceType: { type: string, enum: [ABSENCE, MISSION, EMPLOYMENT, JOB_ROLE, DEMAND, REVISION] }
        sourceId: { type: string }
        entityPath: { type: string }
...
    ApiError:
      type: object
      required: [error, message]
      properties:
        error: { type: string }
        message: { type: string }
        issues: { type: array, items: { type: object } }
        conflicts:
          type: array
          items: { $ref: '#/components/schemas/Conflict' }